requirements.txt
easyocr
torch==2.2.1
langchain
faiss-cpu
langchain_community
sentence-transformers
PyPDF2
python-docx
python-pptx
pydub
moviepy
halo
librosa
SpeechRecognition
cohere
pyttsx3
gradio
langchain_groq
python-bidi==0.4.2